Split SEB pending row processing out of the reconciliation service
The per-row pipeline moves into collaborators along its real seams:
SebPendingRowReconciler matches rows and reports unmatched ones,
SebMatchedRowProcessor validates matched rows, with the consistency
check and the execution upsert with its orphan guard as its own
collaborators, and the RowOutcome enum keeps the matched-but-blocked
path that increments neither counter. The service keeps the loop,
counters, settlement-by-absence detection, and summary alerting, and
its coupling drops from 36 to 24 types. All 57 reconciliation tests
pass unchanged with the real collaborator chain wired in.

Cognitive complexity violations 13 -> 12, PMD violations 113 -> 110,
top-15 class lines 6033 -> 5870.
