+++ metrics/scorecard.json
@@ -1,8 +1,8 @@
 {
-  "branchCoverage": 87.76,
+  "branchCoverage": 87.75,
   "classCouplingViolations": 34,
   "clockViolations": 0,
-  "cognitiveComplexityViolations": 13,
+  "cognitiveComplexityViolations": 12,
   "compilerWarnings": 0,
   "deepNestingViolations": 0,
   "disabledTests": 5,
@@ -26,23 +26,23 @@
   "overMockedTestClasses": 113,
   "overMockedTestShare": 11.2,
   "pitestSlicedModules": 52,
-  "pmdViolations": 113,
+  "pmdViolations": 110,
   "pmdViolationsByRule": {
-    "CognitiveComplexity": 13,
+    "CognitiveComplexity": 12,
     "CouplingBetweenObjects": 34,
-    "CyclomaticComplexity": 61,
+    "CyclomaticComplexity": 59,
     "GodClass": 2,
     "TooManyMethods": 3
   },
-  "slowest10TestClassesSeconds": 113.3,
+  "slowest10TestClassesSeconds": 115.7,
   "springBootTests": 76,
-  "suiteWallTimeSeconds": 222,
-  "testClassesOver10Seconds": 4,
+  "suiteWallTimeSeconds": 197,
+  "testClassesOver10Seconds": 5,
   "testClassesTotal": 1008,
   "testIntegrationShare": 6.6,
   "testSliceShare": 13.7,
   "testUnitShare": 79.7,
-  "top15ClassLines": 6033,
+  "top15ClassLines": 5870,
   "totalPackages": 227,
   "unmarkedPackages": 0
 }
