+++ src/main/java/ee/tuleva/onboarding/investment/transaction/ingest/RowOutcome.java
@@ -1,0 +1,7 @@
+package ee.tuleva.onboarding.investment.transaction.ingest;
+
+enum RowOutcome {
+  MATCHED,
+  UNMATCHED,
+  SKIPPED
+}
