+++ src/main/java/ee/tuleva/onboarding/investment/transaction/ingest/SebPendingTransactionReconciliationService.java
@@ -2,13 +2,7 @@ package ee.tuleva.onboarding.investment.transaction.ingest;
 
 import static ee.tuleva.onboarding.investment.report.ReportProvider.SEB;
 import static ee.tuleva.onboarding.investment.report.ReportType.PENDING_TRANSACTIONS;
-import static ee.tuleva.onboarding.investment.transaction.ingest.ReconciliationAuditRecorder.REASON_FUND_MISMATCH;
-import static ee.tuleva.onboarding.investment.transaction.ingest.ReconciliationAuditRecorder.REASON_ISIN_SIDE_MISMATCH;
-import static ee.tuleva.onboarding.investment.transaction.ingest.ReconciliationAuditRecorder.REASON_MISSING_ISIN;
-import static ee.tuleva.onboarding.investment.transaction.ingest.ReconciliationAuditRecorder.REASON_MISSING_OUR_REF;
-import static java.util.Objects.requireNonNull;
 
-import ee.tuleva.onboarding.fund.TulevaFund;
 import ee.tuleva.onboarding.investment.report.InvestmentReport;
 import ee.tuleva.onboarding.investment.report.InvestmentReportService;
 import ee.tuleva.onboarding.investment.report.SebReportHeaders;
@@ -18,7 +12,6 @@ import ee.tuleva.onboarding.investment.transaction.TransactionExecution;
 import ee.tuleva.onboarding.investment.transaction.TransactionExecutionRepository;
 import ee.tuleva.onboarding.investment.transaction.TransactionOrder;
 import ee.tuleva.onboarding.investment.transaction.TransactionOrderRepository;
-import ee.tuleva.onboarding.investment.transaction.TransactionSettlement;
 import ee.tuleva.onboarding.investment.transaction.TransactionSettlementRepository;
 import ee.tuleva.onboarding.investment.transaction.TransactionSettlementService;
 import java.time.Instant;
@@ -27,7 +20,6 @@ import java.time.ZoneOffset;
 import java.util.Comparator;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
@@ -46,13 +38,8 @@ public class SebPendingTransactionReconciliationService {
   private static final double TRUNCATION_ROW_DROP_RATIO = 0.5;
 
   private final SebPendingTransactionExtractor extractor;
-  private final SebPendingTransactionMatcher matcher;
-  private final SebPendingTransactionComplexMatcher complexMatcher;
   private final QuantityAmountValidator quantityAmountValidator;
-  private final SebClientNameToFundResolver fundResolver;
-  private final ExecutionPriceConsistencyChecker priceConsistencyChecker;
   private final TransactionMatchingPolicy matchingPolicy;
-  private final TransactionExecutionMapper executionMapper;
   private final TransactionExecutionRepository executionRepository;
   private final TransactionOrderRepository orderRepository;
   private final ApplicationEventPublisher eventPublisher;
@@ -60,6 +47,7 @@ public class SebPendingTransactionReconciliationService {
   private final TransactionSettlementRepository settlementRepository;
   private final TransactionSettlementService settlementService;
   private final InvestmentReportService reportService;
+  private final SebPendingRowReconciler rowReconciler;
 
   @Value("${transaction-registry.settlement-check.scan-lookback-days:60}")
   private int scanLookbackDays = 60;
@@ -82,63 +70,13 @@ public class SebPendingTransactionReconciliationService {
     int unmatched = 0;
     Set<Long> presentOrderIds = new HashSet<>();
     for (SebPendingTransactionRow row : rows) {
-      presentOrderIds.addAll(referencedOrderIds(row));
-      Optional<TransactionOrder> orderOpt = matcher.match(row);
-      if (orderOpt.isEmpty()) {
-        orderOpt = matchByBrokerRef(row);
-      }
-      if (orderOpt.isEmpty()) {
-        orderOpt = complexMatcher.match(row, matchingProperties);
-      }
-      if (orderOpt.isEmpty()) {
-        Optional<QuantityAmountMismatchEvent> nearMiss =
-            complexMatcher.findNearMiss(row, matchingProperties);
-        if (nearMiss.isPresent()) {
-          reportMismatch(nearMiss.get().withReportDate(reportDate), row);
-          unmatched++;
-          continue;
-        }
-        log.info(
-            "Unmatched pending transaction: clientRef={}, ourRef={}, isin={}, reportDate={}",
-            row.clientRef(),
-            row.ourRef(),
-            row.isin(),
-            reportDate);
-        auditRecorder.recordUnmatched(row, reportDate);
-        unmatched++;
-        continue;
-      }
-      TransactionOrder order = orderOpt.get();
-      presentOrderIds.add(order.getId());
-      if (!isRowConsistentWithOrder(order, row, reportDate)) {
-        matched++;
-        continue;
-      }
-      Optional<TransactionSettlement> settlement =
-          settlementRepository.findByOrderId(order.getId());
-      if (settlement.isPresent()) {
-        matched++;
-        handleSettledOrderReappearance(order, settlement.get(), row, reportDate);
-        continue;
-      }
-      Optional<QuantityAmountMismatchEvent> blankEconomics =
-          quantityAmountValidator.detectBlankEconomics(order, row, matchingProperties);
-      if (blankEconomics.isPresent()) {
-        reportMismatch(blankEconomics.get().withReportDate(reportDate), row);
-        matched++;
-        continue;
-      }
-      Optional<QuantityAmountMismatchEvent> mismatch =
-          quantityAmountValidator.validateCumulative(
-              order, row, executionRepository.findAllByOrderId(order.getId()), matchingProperties);
-      if (mismatch.isPresent()) {
-        reportMismatch(mismatch.get().withReportDate(reportDate), row);
-        matched++;
-        continue;
-      }
-      if (upsert(row, order, reportDate, asOfDate)) {
-        matched++;
-        checkPriceConsistency(order, reportDate, matchingProperties);
+      RowOutcome outcome =
+          rowReconciler.reconcileRow(
+              row, reportDate, asOfDate, matchingProperties, presentOrderIds);
+      switch (outcome) {
+        case MATCHED -> matched++;
+        case UNMATCHED -> unmatched++;
+        case SKIPPED -> {}
       }
     }
 
@@ -150,170 +88,6 @@ public class SebPendingTransactionReconciliationService {
 
     detectSettlementsByAbsence(
         reportDate, rows.size(), matched, presentOrderIds, extraction.isComplete());
-  }
-
-  private void checkPriceConsistency(
-      TransactionOrder order, LocalDate reportDate, TransactionMatchingProperties properties) {
-    List<TransactionExecution> executions = executionRepository.findAllByOrderId(order.getId());
-    priceConsistencyChecker
-        .check(order, executions, properties.executionPriceConsistencyTolerance())
-        .ifPresent(
-            event -> {
-              log.error(
-                  "Cross-piece price divergence: orderId={}, isin={}, min={}, max={}, spread={},"
-                      + " tolerance={}, reportDate={}",
-                  order.getId(),
-                  order.getInstrumentIsin(),
-                  event.minUnitPrice(),
-                  event.maxUnitPrice(),
-                  event.relativeSpread(),
-                  event.tolerance(),
-                  reportDate);
-              eventPublisher.publishEvent(event.withReportDate(reportDate));
-            });
-  }
-
-  private void reportMismatch(QuantityAmountMismatchEvent event, SebPendingTransactionRow row) {
-    log.info(
-        "Quantity/amount mismatch: clientRef={}, ourRef={}, isin={}, kind={}, expected={},"
-            + " actual={}, reportDate={}",
-        row.clientRef(),
-        row.ourRef(),
-        row.isin(),
-        event.kind(),
-        event.expected(),
-        event.actual(),
-        event.reportDate());
-    auditRecorder.recordQuantityAmountMismatch(event);
-    eventPublisher.publishEvent(event);
-  }
-
-  private Optional<TransactionOrder> matchByBrokerRef(SebPendingTransactionRow row) {
-    if (row.ourRef() == null) {
-      return Optional.empty();
-    }
-    return uniqueExecutionByBrokerRef(row.ourRef())
-        .map(TransactionExecution::getOrderId)
-        .flatMap(orderRepository::findById);
-  }
-
-  private Optional<TransactionExecution> uniqueExecutionByBrokerRef(String brokerRef) {
-    List<TransactionExecution> matches =
-        executionRepository.findAllByBrokerTransactionId(brokerRef);
-    if (matches.size() > 1) {
-      log.error(
-          "Refusing ambiguous broker-ref match: brokerTransactionId={}, executionCount={}",
-          brokerRef,
-          matches.size());
-      return Optional.empty();
-    }
-    return matches.stream().findFirst();
-  }
-
-  private boolean isRowConsistentWithOrder(
-      TransactionOrder order, SebPendingTransactionRow row, LocalDate reportDate) {
-    if (row.ourRef() == null) {
-      log.error(
-          "Matched SEB row has no Our ref, cannot record as a piece: orderId={}, clientRef={},"
-              + " isin={}, reportDate={}",
-          order.getId(),
-          row.clientRef(),
-          row.isin(),
-          reportDate);
-      return rejectInconsistentRow(order, row, REASON_MISSING_OUR_REF, reportDate);
-    }
-    if (row.isin() == null || row.isin().isBlank()) {
-      log.error(
-          "Matched SEB row is missing ISIN, cannot verify instrument: orderId={}, clientRef={},"
-              + " ourRef={}, reportDate={}",
-          order.getId(),
-          row.clientRef(),
-          row.ourRef(),
-          reportDate);
-      return rejectInconsistentRow(order, row, REASON_MISSING_ISIN, reportDate);
-    }
-    if (!row.isin().equals(order.getInstrumentIsin()) || row.side() != order.getTransactionType()) {
-      log.error(
-          "Matched SEB row instrument/side does not match order: orderId={}, orderIsin={},"
-              + " rowIsin={}, orderSide={}, rowSide={}, ourRef={}, reportDate={}",
-          order.getId(),
-          order.getInstrumentIsin(),
-          row.isin(),
-          order.getTransactionType(),
-          row.side(),
-          row.ourRef(),
-          reportDate);
-      return rejectInconsistentRow(order, row, REASON_ISIN_SIDE_MISMATCH, reportDate);
-    }
-    Optional<TulevaFund> rowFund = fundResolver.resolve(row.clientName());
-    if (rowFund.isPresent() && rowFund.get() != order.getFund()) {
-      log.error(
-          "Matched SEB row client name resolves to a different fund than the order: orderId={},"
-              + " orderFund={}, rowFund={}, clientName={}, ourRef={}, reportDate={}",
-          order.getId(),
-          order.getFund(),
-          rowFund.get(),
-          row.clientName(),
-          row.ourRef(),
-          reportDate);
-      return rejectInconsistentRow(order, row, REASON_FUND_MISMATCH, reportDate);
-    }
-    return true;
-  }
-
-  private boolean rejectInconsistentRow(
-      TransactionOrder order, SebPendingTransactionRow row, String reason, LocalDate reportDate) {
-    auditRecorder.recordInconsistentMatchedRow(order, row, reason, reportDate);
-    return false;
-  }
-
-  private boolean upsert(
-      SebPendingTransactionRow row,
-      TransactionOrder order,
-      LocalDate reportDate,
-      LocalDate asOfDate) {
-    if (wouldOrphanExistingExecution(row, order)) {
-      return false;
-    }
-    Optional<TransactionExecution> existing =
-        executionRepository.findByBrokerTransactionId(
-            requireNonNull(row.ourRef(), "Missing ourRef: orderId=" + order.getId()));
-    if (existing.isPresent()) {
-      TransactionExecution execution = existing.get();
-      Map<String, Object> before = executionMapper.mutableFieldsForDeltaAudit(execution);
-      executionMapper.applyTo(execution, row, order);
-      executionRepository.save(execution);
-      Map<String, Object> after = executionMapper.mutableFieldsForDeltaAudit(execution);
-      if (!before.equals(after)) {
-        auditRecorder.recordExecutionUpdated(order, row, reportDate, before, after);
-      }
-    } else {
-      executionRepository.save(executionMapper.toExecution(row, order, asOfDate));
-      auditRecorder.recordExecutionMatched(order, row, reportDate);
-    }
-
-    order.setOrderStatus(OrderStatus.EXECUTED);
-    orderRepository.save(order);
-    return true;
-  }
-
-  private void handleSettledOrderReappearance(
-      TransactionOrder order,
-      TransactionSettlement settlement,
-      SebPendingTransactionRow row,
-      LocalDate reportDate) {
-    if (!reportDate.isAfter(settlement.getReportDate())) {
-      return;
-    }
-    log.error(
-        "Settled order reappeared in pending report: orderId={}, settlementReportDate={},"
-            + " reportDate={}, clientRef={}, ourRef={}",
-        order.getId(),
-        settlement.getReportDate(),
-        reportDate,
-        row.clientRef(),
-        row.ourRef());
-    auditRecorder.recordSettlementReappeared(order, settlement, row, reportDate);
   }
 
   private void detectSettlementsByAbsence(
@@ -426,22 +200,6 @@ public class SebPendingTransactionReconciliationService {
     auditRecorder.recordSettlementDetected(order, reportDate);
   }
 
-  private Set<Long> referencedOrderIds(SebPendingTransactionRow row) {
-    Set<Long> orderIds = new HashSet<>();
-    if (row.ourRef() != null) {
-      uniqueExecutionByBrokerRef(row.ourRef())
-          .map(TransactionExecution::getOrderId)
-          .ifPresent(orderIds::add);
-    }
-    if (row.clientRef() != null) {
-      orderRepository
-          .findByOrderUuid(row.clientRef())
-          .map(TransactionOrder::getId)
-          .ifPresent(orderIds::add);
-    }
-    return orderIds;
-  }
-
   private LocalDate asOfDate(InvestmentReport report) {
     LocalDate asOfDate = SebReportHeaders.asOfDate(report);
     if (asOfDate == null) {
@@ -452,28 +210,5 @@ public class SebPendingTransactionReconciliationService {
       return report.getReportDate();
     }
     return asOfDate;
-  }
-
-  private boolean wouldOrphanExistingExecution(
-      SebPendingTransactionRow row, TransactionOrder order) {
-    if (row.ourRef() == null) {
-      return false;
-    }
-    Optional<TransactionExecution> byBrokerId = uniqueExecutionByBrokerRef(row.ourRef());
-    if (byBrokerId.isEmpty()) {
-      return false;
-    }
-    Long existingOrderId = byBrokerId.get().getOrderId();
-    if (existingOrderId == null || existingOrderId.equals(order.getId())) {
-      return false;
-    }
-    log.warn(
-        "Refusing to re-link execution to different order: brokerTransactionId={},"
-            + " existingOrderId={}, proposedOrderId={}, clientRef={}",
-        row.ourRef(),
-        existingOrderId,
-        order.getId(),
-        row.clientRef());
-    return true;
   }
 }
